Time: 19 ms (28.85%), Space: 45.1 MB (74.35%) - LeetHub

+++ 0437-path-sum-iii/0437-path-sum-iii.java
@@ -19,7 +19,7 @@ class Solution {
             return 0;
         }
         int count = 0;
-        count += findPath(root, (long)targetSum);
+        count += findPath(root, targetSum);
         count += pathSum(root.left, targetSum);
         count += pathSum(root.right, targetSum);
